Move duplicate col_tools in active workflow tables to abstract class
